fix(prs): drop the numbers from the facet menus, and put faces in them
The Author menu listed a number beside every name — 22, 0, 0, 2 — and all of
them were counted from the twenty-five rows in hand while the filter itself
searches the whole repository. So a contributor with three pull requests read
"2", and eleven people read "0" for no reason other than not appearing on page
one. GitHub shows no number in these menus, for exactly this reason.

The number is gone rather than fixed: a true per-option count means one search
per option, which is a dozen round trips to put a digit next to a name nobody
needed a digit for.

Authors and assignees now carry their avatar, as GitHub's do — a face finds a
name in a list of a dozen faster than reading down it does. `Avatar` moves to
its own module to make that possible: the menus could not import it from the
panel, because the panel imports the filter bar, which imports the menus.

Verify: tsc, web 328 tests, vite build; and the menu opened by hand on a repo
with a dozen contributors.
